requirements.txt
numpy>=1.16.1
pandas>=0.24.1
scikit-learn>=0.20.3
scipy>=1.2.1
tensorflow-gpu==1.13.1
matplotlib>=3.0.2
pyEDFlib==0.1.14
mne==0.18.2
wget